+++ pom.xml
@@ -30,7 +30,7 @@
     <dependency>
       <groupId>org.jsoup</groupId>
       <artifactId>jsoup</artifactId>
-      <version>1.14.2</version>
+      <version>1.15.4</version>
     </dependency>
     <!-- Java Servlets API -->
     <dependency>
@@ -67,7 +67,7 @@
     <dependency>
       <groupId>org.jsoup</groupId>
       <artifactId>jsoup</artifactId>
-      <version>1.14.2</version>
+      <version>1.15.4</version>
     </dependency>
 
     <!-- Java Servlets API -->
