+++ src/main/java/com/google/sps/servlets/AddClassServlet.java
@@ -17,31 +17,23 @@ package com.google.sps.servlets;
 import com.google.cloud.datastore.Datastore;
 import com.google.cloud.datastore.DatastoreOptions;
 import com.google.cloud.datastore.Entity;
-import com.google.cloud.datastore.FullEntity;
 import com.google.cloud.datastore.KeyFactory;
 import com.google.cloud.datastore.Query;
 import com.google.cloud.datastore.QueryResults;
 import com.google.cloud.datastore.StringValue;
-import com.google.cloud.datastore.Value;
-import com.google.cloud.datastore.StructuredQuery.OrderBy;
-import com.google.gson.Gson;
-import com.User;
-import com.UserCred;
 
 import org.jsoup.Jsoup;
 import org.jsoup.safety.Safelist;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.List;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-/** Servlet responsible for listing tasks. */
+/** Servlet responsible for adding the course to the user's data*/
 @WebServlet("/add-class")
 public class AddClassServlet extends HttpServlet {
 
@@ -54,11 +46,8 @@ public class AddClassServlet extends HttpServlet {
         Query.newEntityQueryBuilder().setKind("User").build();
     QueryResults<Entity> results = datastore.run(query);
 
-
     List<StringValue> courses = new ArrayList<>();
     List<StringValue> coursesCopy = new ArrayList<>();
-
-
 
     while (results.hasNext()) {
       Entity entity = results.next();
@@ -66,9 +55,6 @@ public class AddClassServlet extends HttpServlet {
       long id = entity.getKey().getId();
       String email = entity.getString("Email");
       if(inputEmail.equals(email) ){
-
-
-        
         courses = entity.getList("courses");
 
         coursesCopy = new ArrayList<StringValue>(courses);
@@ -77,19 +63,12 @@ public class AddClassServlet extends HttpServlet {
           
           KeyFactory keyFactory = datastore.newKeyFactory().setKind("User");
           com.google.cloud.datastore.Key setKey = keyFactory.newKey(id); 
-           Entity set = datastore.get(setKey); 
+          Entity set = datastore.get(setKey); 
 
-           //update the hasImage  
-           set = Entity.newBuilder(set).set("courses", coursesCopy).build(); 
-           datastore.put(set);
+          //update the hasImage  
+          set = Entity.newBuilder(set).set("courses", coursesCopy).build(); 
+          datastore.put(set);
       }
-
-      //UserCred user = new UserCred(email, password);
-      //userList.add(user);
-      
     }
-
-
-
   }
 }
