+++ src/main/webapp/main.css
@@ -1,8 +1,3 @@
-/* font-family: 
- font-family: 'Rooto', sans-serif;
-
- */
-
 *{
   font-family: 'Mukta', sans-serif;
 }
@@ -13,24 +8,20 @@ p{
 
 body{
   margin: 0px;
-    width: 100%;
-    background-color: #F7F9FE;
-}
-
-.app > .main-app {
-
+  width: 100%;
+  background-color: #F7F9FE;
 }
 
 .app{
-    display: flex;
-    flex-direction: row;
+  display: flex;
+  flex-direction: row;
 }
 
 .app > .sidebar {
-    flex-grow: 1;
-    flex-basis: 100%;
-    display: flex;
-    flex-direction: column;
+  flex-grow: 1;
+  flex-basis: 100%;
+  display: flex;
+  flex-direction: column;
 }
 
 
@@ -78,14 +69,12 @@ ul{
 }
 /* Need help: nav bar centering */
 .sidebar ul{
-  /* margin-top: 20px; */
   padding: 2rem;
 
 }
 
 .sidebar ul>li{
   margin-top: 15px;
-  /* position: fixed; */
   height: 50px;
   width: 100%;
 
@@ -105,55 +94,51 @@ ul{
 }
 
 @media all and (min-width: 600px) {
-    .aside { flex: 1 0 0; }
+  .aside { flex: 1 0 0; }
 }
 
 @media all and (min-width: 800px) {
-    .left-col-1 { order: 1; } 
-    .left-col-2   { order: 2; }
-    .left-col-3 { order: 3; }
+  .left-col-1 { order: 1; } 
+  .left-col-2   { order: 2; }
+  .left-col-3 { order: 3; }
 }
 
 /* The Modal (background) */
 
-  
-
 .search-wrapper{
-    display: flex;
-    flex-direction: column;
-    gap: .25rem;
+  display: flex;
+  flex-direction: column;
+  gap: .25rem;
 
 }
 
 input{
-    font-size: 1rem;
+  font-size: 1rem;
 }
 
 .class-crads{
-    display: grid;
-    grid-template-columns: repeat();
-    gap: .25rem;
-    top: 1rem;
+  display: grid;
+  grid-template-columns: repeat();
+  gap: .25rem;
+  top: 1rem;
     
 }
 
 .card{
   z-index: 0;
-    color: #F7F9FE;
-    border: 1px solid rgb(12, 13, 59);
-    background-color: #3a314f;
-    top: 1rem;
-    border-radius: 5px;
-    width: 10rem;
-    height:auto;
+  color: #F7F9FE;
+  border: 1px solid rgb(12, 13, 59);
+  background-color: #3a314f;
+  top: 1rem;
+  border-radius: 5px;
+  width: 10rem;
+  height:auto;
     
 }
 
 #search{
   width: 20rem;
 }
-
-
 
 .modal-container{
   position: fixed;
@@ -197,7 +182,6 @@ input[type="search"]{
 }
 
 .modals{
-  /* margin-top: 2rem; */
   padding: 1rem 2rem;
   background-color: #fcfbff;
   border-radius: 5px;
@@ -221,14 +205,6 @@ input[type="search"]{
 .button-profile{
   background-color: none;
 }
-
-/* .main-element{
-  padding: 20px;
-  margin-bottom: 0;
-} */
-
-.class-cards-wrapper{}
-
 
 .name-card{
   display: flex;
@@ -297,10 +273,11 @@ hr{
   color: aliceblue;
   font-weight: 400;
 }
+
 #class-season{
   font-style: italic;
-  /* font-weight: bold; */
 }
+
 hr{
   border: 0.5px solid#e1d6ff;
   width: 100%;
@@ -314,15 +291,12 @@ hr{
 .class-cards-wrapper{
   display: flex;
   flex-direction: row;
-  /* justify-content:space-evenly; */
   flex-wrap: wrap;
 }
 #hint:hover{
   background-color: #a989ff18;
   border-radius: 5px;
   font-size:x-large;
-  /* color: red; */
-  
 }
